parent pom.xml: maven-compiler.release instead of .source and .target; removed plugin maven-compiler-plugin definition; hw01-maven: pom.xml: build-jdk removed for manifest; HelloOtus: BigIntegerMath conversion to long - removed;

+++ hw01-maven/src/main/java/ru/otus/hw01/HelloOtus.java
@@ -11,7 +11,7 @@ public class HelloOtus {
     public static void main(String... args) throws IOException {
 
         int intReceived = getPositiveInteger();
-        System.out.format("Factorial of %d is %d\n", intReceived, BigIntegerMath.factorial(intReceived).longValue());
+        System.out.format("Factorial of %d is %d\n", intReceived, BigIntegerMath.factorial(intReceived));
 
     }
